// src/actions/types.ts
export enum ActionTypes {
  GET_ARTIST = 'GET_ARTIST',
  GET_ARTISTS = 'GET_ARTISTS',

  GET_WORKS = 'GET_WORKS',
  GET_SONG = 'GET_SONG',
  GET_SONGS = 'GET_SONGS',

  GET_LIVES = 'GET_LIVES',
  GET_LIVE = 'GET_LIVE',

  PREPARE_TOP_PAGE = 'PREPARE/TOP_PAGE',
  PREPARE_ARTIST_PAGE = 'PREPARE/ARTIST_PAGE',
  PREPARE_WORKS_PAGE = 'PREPARE/WORKS_PAGE',
  PREPARE_LIVE_LIST_PAGE = 'PREPARE/LIVE_LIST_PAGE',
  PREPARE_TOUR_PAGE = 'PREPARE/TOUR_PAGE',
  PREPARE_SONG_PAGE = 'PREPARE/SONG_PAGE',
  PREPARE_LIVE_PAGE = 'PREPARE/LIVE_PAGE',
}